Crate ygen_mc

ygen-mc: ygen machine code code playground

With the ygen machine code playground you can play around with assembler instructions and see their opcodes.
It’s suppused to being the ygen variant of the llvm-mc

Usage

Options:

-h, –help Displays help
-v, –version Displays the version
-clr, –color Colorates the ouput
-lex, –show-lexer Displays lexing output

Arguments:

-as=, –assemble-string= The assembly instruction to assemble
-triple=, –triple= The target triple

Functions

main 🔒